requirements.txt
openai-agents
fastapi
uvicorn[standard]
pydantic
python-dotenv
pytest
httpx
jinja2